DEV: Extract build-quote tests (#21217)
They are not related to pretty-text.
